Split speech features normalize
